+++ packages/cli/lib/services/dryrun.service.ts
@@ -197,6 +197,13 @@ class DryRunService {
                 return Promise.resolve([]);
             }
         };
+        // dry-run is not scheduling any tasks so we can safely mock the orchestrator client
+        const orchestratorClient = {
+            execute: () => {
+                return Promise.resolve({}) as any;
+            }
+        };
+
         const logContextGetter = {
             create: () => {
                 return Promise.resolve({}) as any;
@@ -209,6 +216,7 @@ class DryRunService {
         const syncRun = new syncRunService({
             integrationService,
             recordsService,
+            orchestratorClient,
             dryRunService: new DryRunService(environment, true),
             logContextGetter,
             writeToDb: false,
