+++ packages/jobs/Dockerfile
@@ -16,6 +16,7 @@ COPY packages/data-ingestion/ packages/data-ingestion/
 COPY packages/jobs/ packages/jobs/
 COPY packages/logs/ packages/logs/
 COPY packages/runner/ packages/runner/
+COPY packages/orchestrator/ packages/orchestrator/
 COPY package*.json ./
 
 RUN npm pkg delete scripts.prepare
