+++ packages/jobs/package.json
@@ -17,6 +17,7 @@
     "dependencies": {
         "@nangohq/data-ingestion": "file:../data-ingestion",
         "@nangohq/logs": "file:../logs",
+        "@nangohq/nango-orchestrator": "file:../orchestrator",
         "@nangohq/nango-runner": "file:../runner",
         "@nangohq/shared": "file:../shared",
         "@nangohq/records": "file:../records",
