+++ packages/server/lib/controllers/onboarding.controller.ts
@@ -25,7 +25,9 @@ import {
     createActivityLog,
     LogActionEnum,
     analytics,
-    AnalyticsTypes
+    AnalyticsTypes,
+    getOrchestratorUrl,
+    Orchestrator
 } from '@nangohq/shared';
 import type { IncomingPreBuiltFlowConfig } from '@nangohq/shared';
 import { getLogger } from '@nangohq/utils';
@@ -34,6 +36,7 @@ import { logContextGetter } from '@nangohq/logs';
 import { records as recordsService } from '@nangohq/records';
 import type { GetOnboardingStatus } from '@nangohq/types';
 import type { RequestLocals } from '../utils/express.js';
+import { OrchestratorClient } from '@nangohq/nango-orchestrator';
 
 const logger = getLogger('Server.Onboarding');
 
@@ -426,7 +429,8 @@ class OnboardingController {
                     connection: { id: connection.id!, name: connection.connection_id }
                 }
             );
-            const actionResponse = await syncClient.triggerAction({
+            const orchestrator = new Orchestrator(new OrchestratorClient({ baseUrl: getOrchestratorUrl() }));
+            const actionResponse = await orchestrator.triggerAction({
                 connection,
                 actionName: DEMO_ACTION_NAME,
                 input: { title: req.body.title },
