+++ packages/server/package.json
@@ -25,6 +25,7 @@
     "dependencies": {
         "@hapi/boom": "^10.0.1",
         "@nangohq/logs": "file:../logs",
+        "@nangohq/nango-orchestrator": "file:../orchestrator",
         "@nangohq/records": "file:../records",
         "@nangohq/shared": "file:../shared",
         "@nangohq/types": "^0.39.30",
