+++ packages/shared/lib/services/notification/slack.service.ts
@@ -12,6 +12,8 @@ import connectionService from '../connection.service.js';
 import accountService from '../account.service.js';
 import SyncClient from '../../clients/sync.client.js';
 import type { LogContext, LogContextGetter } from '@nangohq/logs';
+import type { OrchestratorClientInterface } from '../../clients/orchestrator.js';
+import { Orchestrator } from '../../clients/orchestrator.js';
 
 const logger = getLogger('SlackService');
 const TABLE = dbNamespace + 'slack_notifications';
@@ -74,12 +76,17 @@ export const generateSlackConnectionId = (accountUUID: string, environmentName: 
  *      - name
  */
 
-class SlackService {
+export class SlackService {
+    private orchestrator: Orchestrator;
     private actionName = 'flow-result-notifier-action';
     private adminConnectionId = process.env['NANGO_ADMIN_CONNECTION_ID'] || 'admin-slack';
     private integrationKey = process.env['NANGO_SLACK_INTEGRATION_KEY'] || 'slack';
     private nangoAdminUUID = process.env['NANGO_ADMIN_UUID'];
     private env = 'prod';
+
+    constructor(orchestratorClient: OrchestratorClientInterface) {
+        this.orchestrator = new Orchestrator(orchestratorClient);
+    }
 
     /**
      * Get Nango Admin Connection
@@ -146,7 +153,7 @@ class SlackService {
             payload.ts = ts;
         }
 
-        const actionResponse = await syncClient.triggerAction<SlackActionResponse>({
+        const actionResponse = await this.orchestrator.triggerAction<SlackActionResponse>({
             connection: nangoAdminConnection,
             actionName: this.actionName,
             input: payload,
@@ -303,7 +310,7 @@ class SlackService {
             return;
         }
 
-        const actionResponse = await syncClient.triggerAction<SlackActionResponse>({
+        const actionResponse = await this.orchestrator.triggerAction<SlackActionResponse>({
             connection: slackConnection as NangoConnection,
             actionName: this.actionName,
             input: payload,
@@ -449,7 +456,7 @@ class SlackService {
             }
         );
 
-        const actionResponse = await syncClient.triggerAction<SlackActionResponse>({
+        const actionResponse = await this.orchestrator.triggerAction<SlackActionResponse>({
             connection: slackConnection as NangoConnection,
             actionName: this.actionName,
             input: payload,
@@ -670,5 +677,3 @@ class SlackService {
             });
     }
 }
-
-export default new SlackService();
