+++ packages/shared/lib/services/sync/run.service.ts
@@ -12,7 +12,7 @@ import localFileService from '../file/local.service.js';
 import { getLastSyncDate, setLastSyncDate } from './sync.service.js';
 import environmentService from '../environment.service.js';
 import accountService from '../account.service.js';
-import slackNotificationService from '../notification/slack.service.js';
+import { SlackService } from '../notification/slack.service.js';
 import webhookService from '../notification/webhook.service.js';
 import { integrationFilesAreRemote, isCloud, getLogger, metrics, stringifyError } from '@nangohq/utils';
 import { getApiUrl, isJsOrTsType } from '../../utils/utils.js';
@@ -25,6 +25,7 @@ import type { Environment } from '../../models/Environment.js';
 import type { LogContext, LogContextGetter } from '@nangohq/logs';
 import type { NangoProps } from '../../sdk/sync.js';
 import type { UpsertSummary } from '@nangohq/records';
+import type { OrchestratorClientInterface } from '../../clients/orchestrator.js';
 
 const logger = getLogger('run.service');
 
@@ -55,6 +56,7 @@ export interface SyncRunConfig {
     integrationService: IntegrationServiceInterface;
     recordsService: RecordsServiceInterface;
     dryRunService?: NangoProps['dryRunService'];
+    orchestratorClient: OrchestratorClientInterface;
     logContextGetter: LogContextGetter;
 
     writeToDb: boolean;
@@ -103,6 +105,7 @@ export default class SyncRun {
     recordsService: RecordsServiceInterface;
     dryRunService?: NangoProps['dryRunService'];
     logContextGetter: LogContextGetter;
+    slackNotificationService: SlackService;
 
     writeToDb: boolean;
     isAction: boolean;
@@ -137,6 +140,7 @@ export default class SyncRun {
         this.integrationService = config.integrationService;
         this.recordsService = config.recordsService;
         this.logContextGetter = config.logContextGetter;
+        this.slackNotificationService = new SlackService(config.orchestratorClient);
         if (config.bigQueryClient) {
             this.bigQueryClient = config.bigQueryClient;
         }
@@ -468,7 +472,7 @@ export default class SyncRun {
                     await this.logCtx?.info(content);
                     await this.logCtx?.success();
 
-                    await slackNotificationService.removeFailingConnection(
+                    await this.slackNotificationService.removeFailingConnection(
                         this.nangoConnection,
                         this.syncName,
                         this.syncType,
@@ -579,7 +583,7 @@ export default class SyncRun {
             // any changes while the sync is running
             if (!this.isWebhook) {
                 await setLastSyncDate(this.syncId as string, syncStartDate);
-                await slackNotificationService.removeFailingConnection(
+                await this.slackNotificationService.removeFailingConnection(
                     this.nangoConnection,
                     this.syncName,
                     this.syncType,
@@ -732,7 +736,7 @@ export default class SyncRun {
 
         if (!this.isWebhook) {
             try {
-                await slackNotificationService.reportFailure(
+                await this.slackNotificationService.reportFailure(
                     this.nangoConnection,
                     this.syncName,
                     this.syncType,
