+++ packages/shared/lib/services/sync/run.service.unit.test.ts
@@ -39,11 +39,18 @@ const recordsService = {
     }
 };
 
+const orchestratorClient = {
+    execute: () => {
+        return Promise.resolve({}) as any;
+    }
+};
+
 describe('SyncRun', () => {
     const dryRunConfig: SyncRunConfig = {
         integrationService: integrationService as unknown as IntegrationServiceInterface,
         recordsService,
-        logContextGetter: logContextGetter,
+        orchestratorClient,
+        logContextGetter,
         writeToDb: false,
         nangoConnection: {
             id: 1,
@@ -62,7 +69,8 @@ describe('SyncRun', () => {
         const config: SyncRunConfig = {
             integrationService: integrationService as unknown as IntegrationServiceInterface,
             recordsService,
-            logContextGetter: logContextGetter,
+            orchestratorClient,
+            logContextGetter,
             writeToDb: true,
             nangoConnection: {
                 id: 1,
