+++ packages/shared/tsconfig.json
@@ -19,6 +19,9 @@
         },
         {
             "path": "../records"
+        },
+        {
+            "path": "../orchestrator"
         }
     ],
     "include": ["lib/**/*", "../utils/lib/vitest.d.ts"]
